fix the build error(warning) in g++ 4.7.2(x64)
building error is following:
ontrol.cc:82:40: error: narrowing conversion of 'supported_compilers.std::vector<_Tp, _Alloc>::size<const Compiler*, std::allocator<const Compiler*> >()' from 'std::vector<const Compiler*>::size_type {aka long unsigned int}' to 'uint32_t {aka unsigned int}' inside { } is ill-formed in C++11 [-Werror=narrowing]

+++ judge_client/client/control.cc
@@ -79,7 +79,7 @@ int ControlMain(const string& queue_address, int queue_port, int port) {
         } else if (command == CMD_INFO) {
             uint32_t buf[128] = {
               static_cast<uint32_t>(port),
-              supported_compilers.size()
+              static_cast<uint32_t>(supported_compilers.size())
             };
             int n = 2;
             for (int i = 0; i < supported_compilers.size(); ++i) {
